@@ -19,9 +19,9 @@
 
 package net.clementraynaud.skoice.listeners.session;
 
-import com.bugsnag.Severity;
 import net.clementraynaud.skoice.Skoice;
 import net.clementraynaud.skoice.bot.BotStatus;
+import net.clementraynaud.skoice.model.minecraft.BasePlayer;
 import net.clementraynaud.skoice.storage.config.ConfigField;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.session.ReadyEvent;
@@ -32,7 +32,6 @@ import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.requests.ErrorResponse;
 import net.dv8tion.jda.api.requests.RestAction;
 import net.kyori.adventure.text.event.HoverEvent;
-import org.bukkit.entity.Player;
 
 import java.time.OffsetDateTime;
 import java.util.function.Consumer;
@@ -48,7 +47,7 @@ public class ReadyListener extends ListenerAdapter {
     @Override
     public void onReady(ReadyEvent event) {
 
-        Player tokenManager = this.plugin.getBot().getTokenManager();
+        BasePlayer tokenManager = this.plugin.getBot().getTokenManager();
 
         event.getJDA().retrieveApplicationInfo().queue(applicationInfo -> {
                     if (applicationInfo.isBotPublic()) {
@@ -78,8 +77,7 @@ public class ReadyListener extends ListenerAdapter {
     private void setDefaultFailure() {
         Consumer<? super Throwable> defaultFailure = RestAction.getDefaultFailure();
         RestAction.setDefaultFailure(throwable -> {
-            if (throwable instanceof ErrorResponseException) {
-                ErrorResponseException error = (ErrorResponseException) throwable;
+            if (throwable instanceof ErrorResponseException error) {
                 if (error.getErrorCode() == ErrorResponse.MISSING_PERMISSIONS.getCode()
                         || error.getErrorCode() == ErrorResponse.MISSING_ACCESS.getCode()
                         || error.getErrorCode() == ErrorResponse.MFA_NOT_ENABLED.getCode()) {
@@ -95,12 +93,11 @@ public class ReadyListener extends ListenerAdapter {
                 return;
             }
 
-            Skoice.analyticManager().getBugsnag().notify(throwable, Severity.ERROR);
             defaultFailure.accept(throwable);
         });
     }
 
-    private void handlePublicBot(Player tokenManager) {
+    private void handlePublicBot(BasePlayer tokenManager) {
         this.plugin.getBot().acknowledgeStatus();
         this.plugin.getConfigYamlFile().remove(ConfigField.TOKEN.toString());
         String botId = this.plugin.getBot().getJDA().getSelfUser().getApplicationId();
@@ -113,7 +110,7 @@ public class ReadyListener extends ListenerAdapter {
         }
 
         if (this.plugin.getConfigYamlFile().getBoolean(ConfigField.TOOLTIPS.toString())) {
-            this.plugin.adventure().sender(tokenManager).sendMessage(this.plugin.getLang().getMessage("chat.configuration.public-bot-interactive", this.plugin.getLang().getComponentMessage("interaction.this-page")
+            tokenManager.sendMessage(this.plugin.getLang().getMessage("chat.configuration.public-bot-interactive", this.plugin.getLang().getComponentMessage("interaction.this-page")
                             .hoverEvent(HoverEvent.showText(this.plugin.getLang().getComponentMessage("interaction.link", "https://discord.com/developers/applications/" + botId + "/bot")))
                             .clickEvent(net.kyori.adventure.text.event.ClickEvent.openUrl("https://discord.com/developers/applications/" + botId + "/bot"))
                     )
@@ -123,7 +120,7 @@ public class ReadyListener extends ListenerAdapter {
         }
     }
 
-    private void handleParsingException(Player tokenManager) {
+    private void handleParsingException(BasePlayer tokenManager) {
         this.plugin.getBot().acknowledgeStatus();
         this.plugin.getConfigYamlFile().remove(ConfigField.TOKEN.toString());
         this.plugin.getBot().getJDA().shutdown();
@@ -135,7 +132,7 @@ public class ReadyListener extends ListenerAdapter {
         }
     }
 
-    private void setup(Player tokenManager) {
+    private void setup(BasePlayer tokenManager) {
         this.plugin.getConfigYamlFile().removeInvalidVoiceChannelId();
 
         this.plugin.getBot().setDefaultAvatar();
