@@ -24,7 +24,7 @@ import net.clementraynaud.skoice.storage.YamlFile;
 import net.clementraynaud.skoice.util.ConfigurationUtil;
 import net.dv8tion.jda.api.entities.channel.concrete.Category;
 import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
-import org.bukkit.configuration.file.YamlConfiguration;
+import org.simpleyaml.configuration.file.YamlConfiguration;
 
 import java.util.Base64;
 import java.util.HashMap;
