@@ -20,8 +20,7 @@
 package net.clementraynaud.skoice.system;
 
 import net.clementraynaud.skoice.Skoice;
-import net.clementraynaud.skoice.util.ThreadUtil;
-import org.bukkit.entity.Player;
+import net.clementraynaud.skoice.model.minecraft.BasePlayer;
 
 import java.util.Collections;
 import java.util.Comparator;
@@ -37,7 +36,6 @@ public class Network {
     private ProximityChannel proximityChannel;
 
     public Network(Skoice plugin, Set<LinkedPlayer> players) {
-        ThreadUtil.ensureNotMainThread();
         this.plugin = plugin;
         this.players.addAll(players);
         Networks.add(this);
@@ -51,7 +49,6 @@ public class Network {
     }
 
     public boolean canPlayerConnect(LinkedPlayer player) {
-        ThreadUtil.ensureNotMainThread();
         if (!player.isStateEligible()) {
             return false;
         }
@@ -62,7 +59,6 @@ public class Network {
     }
 
     public boolean canPlayerStayConnected(LinkedPlayer player) {
-        ThreadUtil.ensureNotMainThread();
         if (!player.isStateEligible()) {
             return false;
         }
@@ -73,7 +69,6 @@ public class Network {
     }
 
     public void splitIfSpread() {
-        ThreadUtil.ensureNotMainThread();
         if (this.size() < 4) {
             return;
         }
@@ -116,43 +111,35 @@ public class Network {
     }
 
     public void engulf(Network network) {
-        ThreadUtil.ensureNotMainThread();
         this.players.addAll(network.players);
         network.players.clear();
     }
 
     public void clear() {
-        ThreadUtil.ensureNotMainThread();
         this.players.clear();
     }
 
     public void add(LinkedPlayer player) {
-        ThreadUtil.ensureNotMainThread();
         this.players.add(player);
     }
 
     public void remove(LinkedPlayer player) {
-        ThreadUtil.ensureNotMainThread();
         this.players.remove(player);
     }
 
-    public void remove(Player player) {
-        ThreadUtil.ensureNotMainThread();
-        this.players.removeIf(p -> p.getBukkitPlayer().equals(player));
+    public void remove(BasePlayer player) {
+        this.players.removeIf(p -> p.getFullPlayer().equals(player));
     }
 
     public boolean contains(LinkedPlayer player) {
-        ThreadUtil.ensureNotMainThread();
         return this.players.contains(player);
     }
 
-    public boolean contains(Player player) {
-        ThreadUtil.ensureNotMainThread();
-        return this.players.stream().anyMatch(p -> p.getBukkitPlayer().equals(player));
+    public boolean contains(BasePlayer player) {
+        return this.players.stream().anyMatch(p -> p.getFullPlayer().equals(player));
     }
 
     public int size() {
-        ThreadUtil.ensureNotMainThread();
         return this.players.size();
     }
 
