@@ -19,7 +19,6 @@
 
 package net.clementraynaud.skoice.tasks;
 
-import com.bugsnag.Severity;
 import net.clementraynaud.skoice.Skoice;
 import net.clementraynaud.skoice.storage.config.ConfigField;
 import net.clementraynaud.skoice.system.LinkedPlayer;
@@ -27,7 +26,6 @@ import net.clementraynaud.skoice.system.Network;
 import net.clementraynaud.skoice.system.Networks;
 import net.clementraynaud.skoice.system.ProximityChannel;
 import net.clementraynaud.skoice.system.ProximityChannels;
-import net.clementraynaud.skoice.util.ThreadUtil;
 import net.dv8tion.jda.api.entities.GuildVoiceState;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
@@ -55,16 +53,15 @@ public class UpdateNetworksTask {
     }
 
     public void start() {
-        this.taskId = this.plugin.getServer().getScheduler().runTaskTimerAsynchronously(
-                this.plugin,
+        this.taskId = this.plugin.getScheduler().runTaskTimerAsynchronously(
                 this::run,
                 0,
-                10
-        ).getTaskId();
+                500
+        );
     }
 
     public void interrupt() {
-        this.plugin.getServer().getScheduler().cancelTask(this.taskId);
+        this.plugin.getScheduler().cancelTask(this.taskId);
 
         for (Pair<String, CompletableFuture<Void>> value : this.awaitingMoves.values()) {
             value.getRight().cancel(true);
@@ -72,10 +69,11 @@ public class UpdateNetworksTask {
     }
 
     private void run() {
-        ThreadUtil.ensureNotMainThread();
         if (!this.lock.tryLock()) {
             return;
         }
+
+        System.out.println("Running UpdateNetworksTask");
 
         try {
             VoiceChannel mainVoiceChannel = this.plugin.getConfigYamlFile().getVoiceChannel();
@@ -136,9 +134,6 @@ public class UpdateNetworksTask {
                     .filter(Objects::nonNull)
                     .count();
             ProximityChannels.clean(possibleUsers);
-        } catch (Throwable throwable) {
-            Skoice.analyticManager().getBugsnag().notify(throwable, Severity.ERROR);
-            throw throwable;
         } finally {
             this.lock.unlock();
         }
@@ -208,7 +203,6 @@ public class UpdateNetworksTask {
     }
 
     private void manageMoves() {
-        ThreadUtil.ensureNotMainThread();
         LinkedPlayer.getOnlineLinkedPlayers().stream()
                 .filter(p -> !p.isInMainVoiceChannel() && !p.isInAnyProximityChannel())
                 .map(p -> this.awaitingMoves.get(p.getDiscordId()))
